Entry point to web app
The default file name for a website's home page (INDEX.HTM is also used). Appropriately named, the home page serves as an index to the main pages on the site, each of which can link to any number of other pages and so on. Typing a URL such as www.computerglossary.com is the same as entering www.computerglossary.com/index.html.